chore: Rename project name
