+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@primno/d365-client",
-  "version": "0.6.1-beta.4",
-  "description": "Dynamics 365 Client for Node.JS",
-  "repository": "github:primno/d365-client",
+  "name": "@primno/dataverse-client",
+  "version": "0.7-beta.0",
+  "description": "Dataverse / Dynamics 365 CE (on-premises) client for Node.JS",
+  "repository": "github:primno/dataverse-client",
   "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
   "types": "dist/index.d.ts",
@@ -55,9 +55,10 @@
     "registry": "https://registry.npmjs.org/"
   },
   "keywords": [
+    "dataverse",
     "dynamics",
     "d365",
-    "webapi",
-    "dataverse"
+    "client",
+    "webapi"
   ]
 }
